Separate VPI wrapper from core functionality
